:recycle: 테스트에 필요한 데이터를 생성하는 메서드들을 common 패키지에서 관리하기 (#14)
