+++ apply/tests/application_test.py
@@ -3,7 +3,7 @@ from django.shortcuts import get_object_or_404, get_list_or_404
 from rest_framework.test import APIClient
 from table.models import Crew, Post, Section, LongSentence, CheckBox, CheckBoxOption, File, User, Administrator
 from .common.test_constants import LONG_SENTENCE_CREATE_REQUEST, CHECKBOX_FILE_CREATE_REQUEST
-
+from common.generator import create_post, create_section, create_long_question, create_checkbox_question, create_option, create_file_question
 
 # 지원서를 생성한다
 class ApplicationCreateTest(TestCase):
@@ -103,49 +103,3 @@ class ApplicationReadTest(TestCase):
         self.assertEqual(len(response.json()[self.section2.section_name]["questions"]), 2)
         self.assertLess(response.json()[self.section2.section_name]["questions"][0]["sequence"], 
                         response.json()[self.section2.section_name]["questions"][1]["sequence"])
-        
-
-# 테스트용 객체를 생성하는 함수들
-def create_post(crew):
-    return Post.objects.create(id=1,
-                            apply_start_date="1000-01-01 00:00:00",
-                            apply_end_date="1000-01-01 00:00:00",
-                            document_result_date="1000-01-01 00:00:00",
-                            has_interview=False,
-                            requirement_target="0",
-                            title="0",
-                            content="0",
-                            membership_fee="0",
-                            crew=crew,
-                            progress="0")
-
-def create_section(id, section_name, post):
-    return Section.objects.create(id=id,
-                                  section_name=section_name,
-                                  post=post,
-                                  description="테스트용 섹션입니다.")
-
-def create_long_question(content, sequence, section):
-    return LongSentence.objects.create(question=content,
-                                       letter_count_limit=500,
-                                       is_essential=True,
-                                       sequence=sequence,
-                                       section=section)
-
-def create_checkbox_question(content, answer_minumum, answer_maximum, sequence, section):
-    return CheckBox.objects.create(question=content,
-                                   is_essential=True,
-                                   answer_minumum=answer_minumum,
-                                   answer_maximum=answer_maximum,
-                                   sequence=sequence,
-                                   section=section)
-
-def create_option(content, check_box):
-    return CheckBoxOption.objects.create(option=content,
-                                         check_box=check_box)
-
-def create_file_question(question, sequence, section):
-    return File.objects.create(question=question,
-                               is_essential=True,
-                               sequence=sequence,
-                               section=section)
